Refactor Items in TextNeuronActivations
